Restructure builder
The main purpose of this commit is to simplify the creation of the
populationsim input files. In doing so, I eliminated the blockgroup level
crosswalk. We now have only region, puma, tract, and taz. We also had
an issue where a single TAZ would appear multiple times in the taz crosswalk. Finally, I ensured that every number that needed to be an integer was coded
as an integer in R.